misc. rename to skillforge in last places
