+++ packages/skillforge-web3/package.json
@@ -20,6 +20,8 @@
     "forge",
     "past3lle",
     "skilltree",
+    "skillforge",
+    "crypto",
     "skills"
   ],
   "author": "pastelle",
