+++ packages/skillforge-widget/package.json
@@ -44,6 +44,8 @@
     "forge",
     "skills",
     "skilltree",
+    "skillforge",
+    "crypto",
     "web3",
     "nft"
   ],
